Refactor `<ImageModal>` to TypeScript
